+++ Mozo/src/main/java/com/incade/poo/mozo/dto/MesaDto.java
@@ -1,5 +1,5 @@
 package com.incade.poo.mozo.dto;
 
-public record MesaDto(Long numero) {
+public record MesaDto(Integer numero) {
 
 }
